pyproject.toml
[tool.poetry]
name = "Sofistik"
version = "0.1.0"
description = "Image creation from sofistik DB"
authors = ["Dmitry Afanasyev <Balshbox@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
